requirements.txt
certifi==2024.8.30
diffusers==0.31.0
huggingface-hub==0.26.2
imageio==2.36.0
imageio-ffmpeg==0.5.1
moviepy==2.1.1
numpy==2.1.3
opencv-python==4.10.0.84
pandas==2.2.3
pillow==10.4.0
pip==24.3.1
requests==2.32.3
safetensors==0.4.5
sentencepiece==0.2.0
setuptools==70.2.0
tokenizers==0.20.4rc0
torch==2.6.0.dev20241126
torchaudio==2.5.0.dev20241126
torchvision==0.20.0.dev20241126
tqdm==4.67.1
transformers==4.46.3
websockets==12.0
